Pertussis

Documentation for Pertussis.

Pertussis.contactmatrix

Pertussis.demorate

Pertussis.pertussisage

Pertussis.phifun

Pertussis.phim

Pertussis.psifun

Pertussis.vaccrate

Pertussis.contactmatrix — Method

contactmatrix(matpath,str)

Read contact matrix mat file by using MAT.jl, and then using DataInterpolations.jl to obtain the contact matrix function.

mathpath is the path of the mat file. str is the key of the dictionary.

Pertussis.demorate — Method

demorate(demopath,str)

Read birth death age migration rate mat file by using MAT.jl, and then using DataInterpolations.jl to obtain the birth rate function.

demopath is the path of the mat file. str is the key of the dictionary.

Pertussis.pertussisage — Method

pertussisage(rhs, StatVar, theta, t)

main model of pertussis. agegroup_str = ["0-1m" "2-11m" "1-4y" "5-9y" "10-14y" "15-19y" "20-24y" 25-29y" "30-34y" "35-39y" "40-44y" "45-49y" "50-54y" "55-59y" "60-64y" "65-69y" "70-74y" "75-79y" "80-84y" "85+y"]

paramstr: $[\sigma_1; \sigma_{2p}; \sigma_{3p};\sigma_{4p}; \sigma^r_scale;\rho; p_1; p_{2p}; p_{3p}; p_{4p}; t_{2}; t_3; \epsilon_p;\epsilon_t]$

Pertussis.phifun — Method

phifun()

coverageadolescent=[0.627;0.627;0.627;0.627;0.627;0.689;0.665;0.677;... 0.699;0.604;0.574;0.65;0.632;0.721;0.741;0.762;0.762]; coverageadult=[0;0;0;0;0;0;0;0;0;0;0.093;0.0965;0.1;0.2155;0.331;0.331;0.34];

Pertussis.phim — Method

phim(t)

maternal booster

Pertussis.psifun — Method

psifun()

rateDTaP2=[5.53;4.21;4.21;3.86;3.59;2.89;4.21;2.36;1.73;1.73;2.36;... 4.21;4.69;2.89;2.89;4.69;4.69]; rateDTaP3=[1.59;1.59;1.59;1.59;1.59;1.43;1.49;1.43;1.29;1.88;1.41;... 1.84;1.87;1.95;1.95;2.66;2.66];

Pertussis.vaccrate — Method

vaccrate(demopath,str)

Read birth death age migration rate mat file by using MAT.jl, and then using DataInterpolations.jl to obtain the birth rate function.

demopath is the path of the mat file. str is the key of the dictionary.